requirements.txt
ccxt>=4.0.0
openai>=1.0.0
python-telegram-bot>=20.0
pandas>=2.0.0
python-dotenv>=1.0.0
